feat：申请友链和友链列表

+++ blog-framework/src/main/java/com/lambkin/blog/model/dto/ArticlePageDto.java
@@ -13,9 +13,9 @@ public class ArticlePageDto {
 
     private String key;
 
-    private Integer current;
+    private Long current;
 
-    private Integer size;
+    private Long size;
 
     private String categoryNo;
 }
